修正严重错误：getFace方法返回错误的face，不过因为face紧凑相连的关系看不出来，在传入边界值时越界程序崩溃。 改动： 1.修改getFace方法传入的参数，没什么影响。 2.将获取点的方式改为常量引用。 3.改变了保留条件，按需更改。

+++ Types.h
@@ -102,7 +102,7 @@ public:
     bool save(const std::string& filename);
 
     void cut();
-    Face* getFace(int faceIndex);
+    Face* getFace(unsigned long long  faceIndex);
     unsigned long long getFaceCount();
     int getMtlIndex(int faceIndex);
